Creating super class
