+++ src/autopas/selectors/tuningStrategy/FullSearch.h
@@ -8,9 +8,9 @@
 
 #include <set>
 #include <sstream>
+#include <utility>
 
-#include "TuningStrategyInterface.h"
-#include "autopas/containers/CompatibleTraversals.h"
+#include "TuningStrategySuperClass.h"
 #include "autopas/selectors/OptimumSelector.h"
 #include "autopas/utils/ExceptionHandler.h"
 
@@ -19,7 +19,7 @@ namespace autopas {
 /**
  * Exhaustive full search of the search space by testing every applicable configuration and then selecting the optimum.
  */
-class FullSearch : public TuningStrategyInterface {
+class FullSearch : public TuningStrategySuperClass {
  public:
   /**
    * Constructor for the FullSearch that generates the search space from the allowed options.
@@ -33,11 +33,9 @@ class FullSearch : public TuningStrategyInterface {
              const std::set<TraversalOption> &allowedTraversalOptions,
              const std::set<DataLayoutOption> &allowedDataLayoutOptions,
              const std::set<Newton3Option> &allowedNewton3Options)
-      : _containerOptions(allowedContainerOptions) {
-    // sets search space and current config
-    populateSearchSpace(allowedContainerOptions, allowedCellSizeFactors, allowedTraversalOptions,
-                        allowedDataLayoutOptions, allowedNewton3Options);
-  }
+      : TuningStrategySuperClass(allowedContainerOptions, allowedCellSizeFactors, allowedTraversalOptions,
+                                 allowedDataLayoutOptions, allowedNewton3Options),
+        _currentConfig(_searchSpace.begin()) {}
 
   /**
    * Constructor for the FullSearch that only contains the given configurations.
@@ -45,17 +43,11 @@ class FullSearch : public TuningStrategyInterface {
    * @param allowedConfigurations Set of configurations AutoPas can choose from.
    */
   explicit FullSearch(std::set<Configuration> allowedConfigurations)
-      : _containerOptions{}, _searchSpace(std::move(allowedConfigurations)), _currentConfig(_searchSpace.begin()) {
-    for (const auto config : _searchSpace) {
-      _containerOptions.insert(config.container);
-    }
-  }
-
-  inline const Configuration &getCurrentConfiguration() const override { return *_currentConfig; }
-
-  inline void removeN3Option(Newton3Option badNewton3Option) override;
+      : TuningStrategySuperClass(std::move(allowedConfigurations)), _currentConfig(_searchSpace.begin()) {}
 
   inline void addEvidence(long time, size_t iteration) override { _traversalTimes[*_currentConfig] = time; }
+
+  inline const Configuration &getCurrentConfiguration() const override { return *_currentConfig; }
 
   inline void reset() override {
     _traversalTimes.clear();
@@ -64,67 +56,14 @@ class FullSearch : public TuningStrategyInterface {
 
   inline bool tune(bool = false) override;
 
-  inline std::set<ContainerOption> getAllowedContainerOptions() const override { return _containerOptions; }
-
-  inline bool searchSpaceIsTrivial() const override { return _searchSpace.size() == 1; }
-
-  inline bool searchSpaceIsEmpty() const override { return _searchSpace.empty(); }
+  inline void removeN3Option(Newton3Option badNewton3Option) override;
 
  private:
-  /**
-   * Fills the search space with the cartesian product of the given options (minus invalid combinations).
-   * @param allowedContainerOptions
-   * @param allowedTraversalOptions
-   * @param allowedDataLayoutOptions
-   * @param allowedNewton3Options
-   */
-  inline void populateSearchSpace(const std::set<ContainerOption> &allowedContainerOptions,
-                                  const std::set<double> &allowedCellSizeFactors,
-                                  const std::set<TraversalOption> &allowedTraversalOptions,
-                                  const std::set<DataLayoutOption> &allowedDataLayoutOptions,
-                                  const std::set<Newton3Option> &allowedNewton3Options);
-
   inline void selectOptimalConfiguration();
 
-  std::set<ContainerOption> _containerOptions;
-  std::set<Configuration> _searchSpace;
   std::set<Configuration>::iterator _currentConfig;
   std::unordered_map<Configuration, size_t, ConfigHash> _traversalTimes;
 };
-
-void FullSearch::populateSearchSpace(const std::set<ContainerOption> &allowedContainerOptions,
-                                     const std::set<double> &allowedCellSizeFactors,
-                                     const std::set<TraversalOption> &allowedTraversalOptions,
-                                     const std::set<DataLayoutOption> &allowedDataLayoutOptions,
-                                     const std::set<Newton3Option> &allowedNewton3Options) {
-  // generate all potential configs
-  for (const auto &containerOption : allowedContainerOptions) {
-    // get all traversals of the container and restrict them to the allowed ones
-    const std::set<TraversalOption> &allContainerTraversals =
-        compatibleTraversals::allCompatibleTraversals(containerOption);
-    std::set<TraversalOption> allowedAndApplicable;
-    std::set_intersection(allowedTraversalOptions.begin(), allowedTraversalOptions.end(),
-                          allContainerTraversals.begin(), allContainerTraversals.end(),
-                          std::inserter(allowedAndApplicable, allowedAndApplicable.begin()));
-
-    for (const auto &cellSizeFactor : allowedCellSizeFactors)
-      for (const auto &traversalOption : allowedAndApplicable) {
-        for (const auto &dataLayoutOption : allowedDataLayoutOptions) {
-          for (const auto &newton3Option : allowedNewton3Options) {
-            _searchSpace.emplace(containerOption, cellSizeFactor, traversalOption, dataLayoutOption, newton3Option);
-          }
-        }
-      }
-  }
-
-  AutoPasLog(debug, "Points in search space: {}", _searchSpace.size());
-
-  if (_searchSpace.empty()) {
-    autopas::utils::ExceptionHandler::exception("FullSearch: No valid configurations could be created.");
-  }
-
-  _currentConfig = _searchSpace.begin();
-}
 
 bool FullSearch::tune(bool) {
   // repeat as long as traversals are not applicable or we run out of configs
@@ -184,5 +123,4 @@ void FullSearch::removeN3Option(Newton3Option badNewton3Option) {
         "Removing all configurations with Newton 3 {} caused the search space to be empty!", badNewton3Option);
   }
 }
-
 }  // namespace autopas
